@@ -11,7 +11,7 @@ import OSLog
 import Models
 import WidgetFeature
 
-public protocol ScreenshotImporting {
+public protocol ScreenshotImporting: Sendable {
     func screenshots(from source: PhotoSource) async throws -> [UIScreenshot]
 }
 
@@ -39,12 +39,20 @@ public struct ScreenshotImporter: ScreenshotImporting {
             switch deepLink {
             case .latestScreenshot:
                 logger.info("Fetching latest screenshot.")
-                let screenshot = try await imageManager.latestScreenshot(from: url)
+                let (screenshot, _) = try await imageManager.latestScreenshot(using: .url(url))
                 logger.info("Successfully fetched latest screenshot.")
                 return [screenshot]
             case .multipleScreenshots:
                 logger.info("Fetching multiple screenshots.")
-                let screenshots = try await imageManager.multipleScreenshots(from: url)
+                let durationString = try deepLinkManager.deepLinkValue(from: url)
+                guard
+                    let duration = Int(durationString),
+                    let option = DurationWidgetOption(rawValue: duration)
+                else {
+                    throw DeepLinkManager.DeepLinkManagerError.badDeepLinkURL
+                }
+                
+                let screenshots = try await imageManager.multipleScreenshots(for: option)
                 logger.info("Retrieved (\(screenshots.count, privacy: .public)) screenshots.")
                 return screenshots
             }
@@ -68,6 +76,39 @@ public struct ScreenshotImporter: ScreenshotImporting {
         case .existingScreenshots(let existing):
             logger.info("Using existing screenshots (\(existing.count, privacy: .public)).")
             return existing
+        case .controlCenter(let id):
+            let imageManager = ImageManager()
+
+            switch id {
+            case 0:
+                let (screenshot, _) = try await imageManager.latestScreenshot(using: .latest)
+                return [screenshot]
+            case 1...4:
+                guard let option = DurationWidgetOption(rawValue: id - 1) else {
+                    throw Error.invalidOption
+                }
+                
+                return try await imageManager.multipleScreenshots(for: option)
+            default:
+                throw Error.invalidID
+            }
         }
+    }
+    
+    // MARK: - Errors
+    
+    public struct Error: LocalizedError {
+        public let errorDescription: String?
+        public let recoverySuggestion: String?
+                
+        public static let invalidOption = Self(
+            errorDescription: "Invalid Option",
+            recoverySuggestion: "The option that was passed in is invalid"
+        )
+        
+        public static let invalidID = Self(
+            errorDescription: "Invalid ID",
+            recoverySuggestion: "The id that was passed in is invalid"
+        )
     }
 }
